📦️ (line_comments.py): Abstract line_comments.py ♻️ (is_empty.py): Move remove_line_comment(s) functions to line_comments.py new file:   line_comments.py modified:   is_empty.py
