+++ is_empty.py
@@ -1,4 +1,4 @@
-from utils import first
+from line_comments import remove_line_comments
 
 
 def is_empty_file(
@@ -43,26 +43,6 @@ def is_empty_text(
         return text.isspace()
 
     return text == ''
-
-
-def remove_line_comments(
-    text: str,
-    line_comment_char: str
-):
-    new_text = []
-
-    for line in text.split('\n'):
-        new_line = remove_line_comment(line, line_comment_char)
-        new_text.append(new_line)
-
-    return '\n'.join(new_text)
-
-
-def remove_line_comment(
-    line: str,
-    line_comment_char: str
-):
-    return first(line.split(line_comment_char))
 
 
 def remove_block_comments(
